tweaks to silence/fix clang compiler warnings

+++ src/thd_niftiwrite.c
@@ -653,7 +653,7 @@ void nifti_set_afni_extension( THD_3dim_dataset *dset , nifti_image *nim )
 
    /* n2   10 Jul, 2015 [rickr] */
    sprintf(buf,"%ld,%ld,%ld,%ld,%ld,%d" ,
-           nim->nx, nim->ny, nim->nz, nim->nt, nim->nu, nim->datatype ) ;
+           (int)nim->nx, (int)nim->ny, (int)nim->nz, (int)nim->nt, (int)nim->nu, (int)nim->datatype ) ;
    NI_set_attribute( ngr , "NIfTI_nums" , buf ) ;
 
    /** now, scan attribute elements in the group, and mark some
